backend/requirements.txt
nltk
scikit-learn
Flask
Flask-Cors
vosk
python-dotenv
requests
pydub
pytesseract
Pillow
